backend/requirements.txt
# FastAPI и асинхронные библиотеки
fastapi>=0.106.0
uvicorn>=0.24.0
websockets>=11.0.3
asyncio>=3.4.3
httpx>=0.25.0
python-multipart>=0.0.6

# Langchain и интеграции
langchain>=0.0.352
langchain-community>=0.0.13
langchain-core>=0.1.17
langchain-openai>=0.0.5

# Работа с векторами
faiss-cpu>=1.7.4
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Работа с документами
python-docx>=1.0.1
pypdf>=3.17.1
unstructured>=0.10.30
markdown>=3.5.1
pdfminer.six>=20220524
tabulate>=0.9.0
pandas>=2.1.3
openpyxl>=3.1.2

# Инструменты для работы с файлами
chardet>=5.2.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Хранение данных
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0
redis>=5.0.1